#9 refactor: remove unnecessary class and line
