+++ halu5071/app/src/main/java/io/moatwel/github/domain/usecase/TokenUseCase.kt
@@ -2,6 +2,6 @@ package io.moatwel.github.domain.usecase
 
 import io.moatwel.github.domain.repository.AuthDataRepository
 
-class TokenUsecase(
+class TokenUseCase(
   val tokenRepository: AuthDataRepository
 )
