+++ halu5071/app/src/main/java/io/moatwel/github/domain/usecase/UserUseCase.kt
@@ -2,7 +2,7 @@
  *
  *  GitHub-Client
  *
- *  MainActivityComponent.kt
+ *  UserUsecase.kt
  *
  *  Copyright 2018 moatwel.io
  *  author : halu5071 (Yasunori Horii)
@@ -21,14 +21,18 @@
  *
  */
 
-package io.moatwel.github.presentation.di.activity
+package io.moatwel.github.domain.usecase
 
-import dagger.Subcomponent
-import dagger.android.AndroidInjector
-import io.moatwel.github.presentation.view.activity.MainActivity
+import io.moatwel.github.domain.entity.User
+import io.moatwel.github.domain.repository.UserRepository
+import io.reactivex.Observable
+import javax.inject.Inject
 
-@Subcomponent
-interface MainActivityComponent : AndroidInjector<MainActivity> {
-  @Subcomponent.Builder
-  abstract class Builder : AndroidInjector.Builder<MainActivity>()
+class UserUseCase @Inject constructor(
+  private val userRepository: UserRepository
+) {
+
+  fun get(): Observable<User> {
+    return userRepository.get()
+  }
 }
