+++ halu5071/app/src/main/java/io/moatwel/github/presentation/di/AppComponent.kt
@@ -14,7 +14,7 @@ import javax.inject.Singleton
   NetworkModule::class,
   RepositoryModule::class,
   DataSourceModule::class,
-  UsecaseModule::class,
+  UseCaseModule::class,
   MainActivityBuilder::class
 ])
 interface AppComponent {
@@ -25,7 +25,7 @@ interface AppComponent {
     fun application(app: Application): Builder
 
     fun addNetworkModule(networkModule: NetworkModule): Builder
-    fun addUsecaseModule(usecaseModule: UsecaseModule): Builder
+    fun addUsecaseModule(usecaseModule: UseCaseModule): Builder
     fun addRepositoryModule(repositoryModule: RepositoryModule): Builder
     fun addDatasourceModule(datasourceModule: DataSourceModule): Builder
     fun build(): AppComponent
